Bayesian_SegNet/requirements.txt
jupyter>=1.0.0
Keras>=2.2.2
matplotlib>=2.2.3
moviepy>=0.2.3.5
numpy>=1.14.5
pandas>=0.23.4
Pillow==5.0.0
scikit-image>=0.14.0
scikit-learn>=0.19.0
tabulate>=0.8.2
tensorflow-gpu>=1.11.0
tqdm>=4.23.4